requirements.txt
corva-sdk==1.9.0
pytest==7.0.1
requests-mock==1.11.0
requests==2.29